+++ App.js
@@ -6,8 +6,9 @@ import {
   Dimensions,
   Text,
   TouchableOpacity,
+  View,
 } from "react-native";
-import { createStackNavigator } from "@react-navigation/stack";
+
 import Phase from "./components/Phase";
 import phasesData from "./data/phases.json";
 import Button from "./components/Button";
@@ -15,16 +16,23 @@ import {
   widthPercentageToDP as wp,
   heightPercentageToDP as hp,
 } from "react-native-responsive-screen";
-const Stack = createStackNavigator();
+
 const screenHeight = Dimensions.get("window").height;
 const screenWidth = Dimensions.get("window").width;
 
 const App = () => {
   const [phases, setPhases] = useState(phasesData);
 
+  // ADD EACH PHASE
   const addPhase = () => {
-    console.log("hlello");
-    setPhases((prev) => [...phases, prev]);
+    const newPhase = { id: phases.length + 1 };
+    setPhases((prev) => [...prev, newPhase]);
+  };
+
+  // REMOVE EACH PHASE
+  const deletePhase = (id) => {
+    const filteredPhases = phases.filter((phase) => phase.id !== id);
+    setPhases(filteredPhases);
   };
 
   return (
@@ -35,9 +43,15 @@ const App = () => {
     >
       <Text style={styles.title}>MW - TODO</Text>
       <ScrollView horizontal style={styles.container}>
-        {phases.map((phase) => (
-          <Phase key={phase.id} phase={phase} />
+        {phases.map((phase, index) => (
+          <View key={index}>
+            <TouchableOpacity onPress={() => deletePhase(phase.id)}>
+              <Text>Delete Phase</Text>
+            </TouchableOpacity>
+            <Phase key={phase.id} phase={phase} />
+          </View>
         ))}
+
         <TouchableOpacity style={styles.buttonContainer} onPress={addPhase}>
           <Text style={styles.buttonText}>+ Add phase</Text>
         </TouchableOpacity>
@@ -56,7 +70,6 @@ const styles = StyleSheet.create({
   },
   container: {
     margin: 5,
-    gap: 10,
     marginRight: 16,
   },
   backgroundImage: {
@@ -77,5 +90,4 @@ const styles = StyleSheet.create({
     marginTop: 5,
   },
 });
-
 export default App;