+++ CHANGES.md
@@ -5,3 +5,5 @@
 - added useState to store the title and description value
 - Added TextInput to show the text and edit the text instead of Text component
 - Added new library react-native-responsive-screen
+- changes the button style
+-